% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{add_log_modifications}
\alias{add_log_modifications}
\title{modify plots to log scale (helper function - do not export)}
\usage{
add_log_modifications(p)
}
\description{
modify plots to log scale (helper function - do not export)
}
