% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{blqhist}
\alias{blqhist}
\title{Summary of BLQ data}
\usage{
blqhist(df, facet_on = NULL, xlab = "Time", ylab = "count")
}
\arguments{
\item{df}{input data frame}

\item{facet_on}{facet variable}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}
}
\value{
A \code{ggplot} object
}
\description{
Graphical Summary of Data Below Limit of Quantitation
}
\section{Required variables}{

\subsection{BLQ}{
Binary column equal to 1 for records below limit of quantitation, 0 otherwise
}
\subsection{EVID}{
Binary column equal to 1 for dosing records, 0 otherwise
}
\subsection{TIME}{
Time variable
}
}

