% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{correlation.matrix}
\alias{correlation.matrix}
\title{Creates a correlation matrix with covariates}
\usage{
correlation.matrix(df)
}
\arguments{
\item{df}{input data frame}
}
\description{
Creates a correlation matrix with covariates
}
\section{another section}{
NA
}

\examples{
dat <- df \%>\% dplyr::distinct(USUBJID, .keep_all=TRUE) \%>\% dplyr::select(AGE, BWT, BHT, BBMI, SEXF)
p <- correlation.matrix(dat)
print(p)
}
