% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{mean_plot}
\alias{mean_plot}
\title{Creates a mean plot object with LOESS curve representing mean}
\usage{
mean_plot(
  df,
  time = "TIME",
  dv = "DV",
  scale = "linear",
  xval = "Actual time since first dose",
  xunit = "hr",
  yval = "Drug concentration",
  yunit = "ng/mL",
  facet_on = "NOFACET"
)
}
\arguments{
\item{df}{input data frame}

\item{time}{column variable representing time, typically "TIME"}

\item{dv}{column variable representing dependent variable, typically "DV"}

\item{scale}{either "linear" or "semilog"}

\item{xval}{string for name of x value}

\item{xunit}{string for unit of x value}

\item{yval}{string for name of y value}

\item{yunit}{string for unit of y value}

\item{facet_on}{facet variable, e.g. "RACE", "SEX", etc}
}
\description{
Creates a mean plot object with LOESS curve representing mean
}
