% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{peek}
\alias{peek}
\title{Summarize data types}
\usage{
peek(df)
}
\arguments{
\item{df}{input data frame}
}
\value{
A \code{flextable} object
}
\description{
Summarize dataset column names and data types
}
\note{
The \code{flextable} package needs to be installed for this to work.
}
