% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppk.df}
\alias{ppk.df}
\title{Example population pharmacokinetic (POPPK) dataset}
\format{
An object of class \code{data.frame} with 5310 rows and 49 columns.
}
\usage{
ppk.df
}
\description{
Contains columns routinely included in POPPK analysis
}
\examples{
data(ppk.df) # Lazy loading
}
\keyword{datasets}
