% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{stackbar}
\alias{stackbar}
\title{Stacked bar plots}
\usage{
stackbar(df, gval, sval, xlab = "", ylab = "Count", fill.lab = "")
}
\arguments{
\item{df}{input data frame}

\item{gval}{column to group on (x-axis tick labels)}

\item{sval}{column to summarize on (stacks)}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{fill.lab}{legend title}
}
\value{
A \code{ggplot} object
}
\description{
Create stacked barplots to tally variable of interest
}
