% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{new_tab}
\alias{new_tab}
\title{Creates a new Rmarkdown html tab.}
\usage{
new_tab(title = "Tab", header = 2, nonumber = T, subtabs = F, pills = F)
}
\arguments{
\item{title}{Tab title}

\item{header}{An int indicating Rmd html header level (e.g. 3 means "###")}

\item{subtabs}{A boolean for whether tab has subtabs below it or not}

\item{pills}{A boolean for whether tab should be pill-shaped or not}

\item{numbered}{A boolean for whether tab should be numbered or not}
}
\description{
Make sure that Rmd chunk option results='asis' is present.
}
\examples{
p <- ggplot(x, y, dat=somedata)
new_tab(title="My Plot", header=3)
print(p)
}
